% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ic}
\alias{ic}
\title{Construct an interval-censored covariate for formulas}
\usage{
ic(zl, zr, name = "z")
}
\arguments{
\item{zl}{Numeric vector of left endpoints.}

\item{zr}{Numeric vector of right endpoints.}

\item{name}{(optional) Name of the variable; defaults to \emph{"z"}.}
}
\description{
Helper used inside model formulas to bundle left/right endpoints of an
interval-censored covariate.
}
