% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icglm_helpers.R
\name{icglm-methods}
\alias{icglm-methods}
\alias{coef.icglm}
\alias{dispersion.icglm}
\alias{vcov.icglm}
\alias{summary.icglm}
\alias{print.summary.icglm}
\title{Helpers for \code{icglm} objects}
\usage{
\method{coef}{icglm}(object, ...)

dispersion.icglm(object, ...)

\method{vcov}{icglm}(object, ...)

\method{summary}{icglm}(object, ...)

\method{print}{summary.icglm}(
  x,
  digits = max(3, getOption("digits") - 3),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{object}{An object of class \code{icglm}.}

\item{...}{Further arguments passed to or from methods.}

\item{x}{An object of class \code{summary.icglm}.}

\item{digits}{Minimum number of significant digits to print.}

\item{signif.stars}{Logical; if \code{TRUE}, print significance stars.}
}
\description{
Minimal S3 methods to work with model fits of class \code{icglm},
analogous to \code{glm} methods
}
\examples{
fit <- icglm(RNA ~ age + ic(zl, zr, "time"),
             family = Gamma("log"), data = actg359,
             Lin = TRUE, Rin = FALSE)
summary(fit)
coef(fit)
vcov(fit)

}
\seealso{
\code{\link{icglm}}, \code{\link[stats]{glm}}, \code{\link[stats]{coef}}, \code{\link[stats]{vcov}},
\code{\link[stats]{printCoefmat}}
}
\concept{icglm}
