% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icglm_residuals.r
\name{ires}
\alias{ires}
\title{Pearson, deviance and quantile residuals for a GLM with an interval-censored covariate}
\usage{
ires(theta, y, zl, zr, X, famly, type = "all")
}
\arguments{
\item{theta}{vector of parameter estimates \eqn{(\hat\beta,\hat\gamma,\hat\phi)};
\eqn{\hat\phi} is only used for quantile residuals.}

\item{y}{Response vector.}

\item{zl, zr}{Numeric vectors with left and right endpoints of the
observed intervals for the interval-censored covariate.}

\item{X}{Design matrix for the non-censored covariates.}

\item{famly}{A family object returned by \code{\link{icglm}}.}

\item{type}{Character vector of residual types (\code{"pearson"}, \code{"deviance"},
\code{"quantile"}); \code{"all"} (default).}
}
\value{
A list.
}
\description{
Computes residuals for a GLM
\eqn{\mu=\mathrm{linkinv}(X^\top\beta + Z\gamma)},
where \code{Z} is only known to lie between \code{zl} and \code{zr}. Residuals are returned
in two ways: (i) as functions \eqn{r(z)} for a supplied value of \code{z} (via \code{fpear},
\code{fdev}, \code{fquant}), and (ii) as interval-censored residuals with endpoints
\code{rl} and \code{rr} (via \code{icpear}, \code{icdev}, \code{icquant}).
}
\details{
Quantile residuals are based on the Probability Integral Transform (PIT).
For continuous families the PIT \eqn{u=F(y\mid\hat\mu)} is uniquely determined by \eqn{\hat\mu},
and the quantile residual is \eqn{r=\Phi^{-1}(u)}, where \eqn{\Phi} is the standard normal cdf.
When \eqn{\hat\mu} is interval-censored with endpoints \code{mul} and \code{mur}, the PIT
is between \eqn{F(y\mid \mathrm{mur})} and \eqn{F(y\mid \mathrm{mul})}.
Function \code{fquant} computes \eqn{r=\Phi^{-1}(u)} at a supplied \eqn{\hat\mu},
while \code{icquant} corresponds to the residual interval induced by \code{mul} and \code{mur}.

For discrete families (\code{binomial} and \code{poisson}), given \eqn{\hat\mu} the PIT is an interval
\eqn{u\in[F(y^-\mid\hat\mu),\,F(y\mid\hat\mu)]} due to jumps in the cdf. With interval-censored \eqn{\hat\mu},
these endpoints are combined with \code{mul} and \code{mur} to form \code{icquant}.
In this case, \code{fquant} returns \emph{randomized} quantile residuals by sampling
\eqn{u^* \sim \mathrm{Unif}(F(y^-\mid\hat\mu),\,F(y\mid\hat\mu))} and returning \eqn{\Phi^{-1}(u^*)}.
}
\examples{
data(actg359)

fit2 <- icglm(RNA ~ age + ic(zl, zr, "waitime"), family = Gamma("log"),
            data = actg359, Lin = TRUE, Rin = FALSE)
out <- ires(fit2$thetaHat, actg359$RNA, actg359$zl, actg359$zr,
            model.matrix(RNA~age, actg359), fit2$family)

}
\seealso{
\code{\link{icglm}}
}
