% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icplots.R
\name{plot.wHat}
\alias{plot.wHat}
\title{Plot augmented Turnbull estimator(s)}
\usage{
\method{plot}{wHat}(lwHat, col = 1:length(lwHat), nleg = "", gini = NULL)
}
\arguments{
\item{lwHat}{List of augmented Turnbull estimators returned by \code{icglm}.}

\item{col}{Vector of colors (one per element in \code{lwHat}).}

\item{nleg}{Character string used as legend title.}

\item{gini}{Optional \pkg{ggplot2} object to add layers to; if \code{NULL},
a new plot is created.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
Plots the augmented Turnbull estimator returned by \code{icglm}. Accepts a list
\code{lwHat} of one or more estimators for the same censored covariate and overlays them.
}
\examples{
data(actg359)
linmod <- icglm(logRNA ~ age + ic(zl,zr,"waitime"), family = gaussian,
                data = actg359, Lin = F, Rin = T)
gammod <- icglm(RNA ~ age + ic(zl,zr,"waitime"), family = Gamma("log"),
                data = actg359, Lin = F, Rin = T)
plot.wHat(list("LM" = linmod$wHat, "GLM" = gammod$wHat))

}
