% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval2.R
\name{Aintmap2}
\alias{Aintmap2}
\title{Construct augmented Turnbull intervals and incidence matrix}
\usage{
Aintmap2(
  L,
  R,
  Lin = TRUE,
  Rin = TRUE,
  xmin = 0,
  xmax = max(R),
  keepall = FALSE
)
}
\arguments{
\item{L, R}{Numeric vectors of left/right endpoints of the observed intervals.}

\item{Lin, Rin}{Logical scalar or length-\eqn{n} vectors indicating whether
the left/right endpoints are included.}

\item{xmin, xmax}{Numeric bounds of the support of \eqn{Z}.
(defaults: \code{xmin = 0}, \code{xmax = max(R)}).}

\item{keepall}{Logical; if \code{FALSE} (default), remove partition intervals
with zero contribution (columns with all zeros).}
}
\description{
Construct augmented Turnbull intervals and incidence matrix
}
\details{
This function adapts function Aintmap from the \CRANpkg{interval} package
to handle augmented Turnbull intervals.
}
\seealso{
\code{\link[interval]{Aintmap}}
}
