% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval2.R
\name{icfit2}
\alias{icfit2}
\title{Turnbull's estimator along the augmented Turnbull intervals}
\usage{
icfit2(L, R, Lin = T, Rin = T, keepall = FALSE, accu = 1e-10)
}
\description{
Turnbull's estimator along the augmented Turnbull intervals
}
\details{
This function is adapted from the \CRANpkg{interval} package to operate on
augmented Turnbull partitions.
}
\seealso{
\code{\link[interval]{icfit}}
}
