% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icglm_helpers.R
\name{icglm-methods}
\alias{icglm-methods}
\alias{coef.icglm}
\alias{dispersion.icglm}
\alias{vcov.icglm}
\alias{summary.icglm}
\alias{print.summary.icglm}
\title{Helpers for \code{icglm} objects}
\usage{
\method{coef}{icglm}(object, ...)

dispersion.icglm(object, ...)

\method{vcov}{icglm}(object, ...)

\method{summary}{icglm}(object, ...)

\method{print}{summary.icglm}(
  x,
  digits = max(3, getOption("digits") - 3),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\description{
Minimal S3 methods to work with model fits of class \code{icglm},
analogous to \code{glm} methods
}
\examples{
fit <- icglm(RNA ~ age + ic(zl, zr, "time"),
             family = Gamma("log"), data = actg359,
             Lin = FALSE, Rin = TRUE)
summary(fit)
coef(fit)
vcov(fit)

}
\seealso{
\code{\link[stats]{coef.glm}}, \code{\link[stats]{vcov.glm}},
\code{\link[stats]{summary.glm}}
}
\concept{icglm}
