% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-toy-icglm-diagnostics.R
\docType{data}
\name{toy_icglm_diagnostics}
\alias{toy_icglm_diagnostics}
\alias{toy_resnorm}
\alias{toy_linearity}
\alias{toy_link}
\title{Toy examples for model diagnostics in a GLM with an interval-censored covariate}
\format{
An object of class \code{list} of length 3.

An object of class \code{list} of length 3.

An object of class \code{list} of length 3.
}
\usage{
toy_resnorm

toy_linearity

toy_link
}
\description{
These objects are simulated toy examples shipped with the package to
illustrate diagnostic checks for \code{icglm} models without requiring long
model fitting in \code{@examples}.
}
\details{
The toy examples correspond to those in:
Gómez Melis, G., Toloba, A., and Langohr, K. (forthcoming).
"Residual Diagnostics for Generalized Linear Models with Interval-Censored Covariates."
In: Du, M., Chen, D.-G., Jin, Z., and Sun, J. (eds.),
\emph{Next-Gen Lifetime Data Analysis: Emerging Innovations and Applications}.
Springer.

Each object is a named list that contains a dataset plus pre-fitted \code{icglm} models.
}
\keyword{datasets}
